requirements.txt
aioimaplib==2.0.1
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asyncpg==0.30.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.14
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
IMAPClient==3.0.1
jiter==0.10.0
Levenshtein==0.27.1
openai==1.98.0
opentimestamps==0.4.5
opentimestamps-client==0.7.2
playwright==1.53.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PySocks==1.7.1
python-bitcoinlib==0.12.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-whois==0.9.5
RapidFuzz==3.13.0
requests==2.32.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
validate-email-address==1

